Fix bug for Bluearc/HDS where bk chmod would dump core when run outside a repository.
bk: 56f30cb107eYLhOnSrofqBdH3x9mzA

+++ src/bk.sh
@@ -728,6 +728,7 @@ _chmod() {		# /* doc 2.0 */
 	MODE=$1
 	shift
 	ROOT=`bk root -R`
+	test $? -eq 0 || exit 1
 	rm -f "$ROOT/BitKeeper/tmp/err$$"
 	bk gfiles ${1+"$@"} | while read i
 	do	test "X`bk gfiles -c $i`" = X || {
